refactor and tidy
